Run all database operations on the same thread
When using the Bukkit scheduler, it's possible that two database operations are run at the same time on different threads, resulting in one of them failing due to the database being locked.

To avoid this, each DBFile now has its own ExecutorService with one thread that all database operations must be run through.

Closes #23.