Manually calculate which players should see indicators
This is originally how it used to work, but in 91caf74e (2022) it was
rewritten to use Minecraft's entity trackers. This has caused numerous
issues since, ranging from ghost indicators to compilation errors
between versions. This finally undoes that misguided optimization and
gives us a system that'll be easier to maintain, and more reliable.